langchain_core_tutorials/requirements.txt
# LangChain Core 核心依赖
langchain-core>=0.3.0
langchain-openai>=0.3.0

# 工具
python-dotenv>=1.0.0
pydantic>=2.0.0
